requirements.txt
click==8.1.3
Flask==2.2.3
Flask-Login==0.6.3
gspread==6.1.2
gunicorn==23.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.2
oauth2client==4.1.3
pickleshare==0.7.5
python-dateutil==2.8.2
scikit-learn==1.4.2
xgboost==2.1.1
Werkzeug==2.2.3
